+++ automated-tests/src/dali/utc-Dali-Layer.cpp
@@ -516,7 +516,7 @@ int UtcDaliLayerPropertyIndices(void)
   Property::IndexContainer indices;
   layer.GetPropertyIndices( indices );
   DALI_TEST_CHECK( indices.Size() > basicActor.GetPropertyCount() );
-  DALI_TEST_EQUALS( indices.Size(), layer.GetPropertyCount(), TEST_LOCATION );
+  DALI_TEST_EQUALS( (unsigned int)indices.Size(), layer.GetPropertyCount(), TEST_LOCATION );
   END_TEST;
 }
 
@@ -563,7 +563,7 @@ int UtcDaliLayerClippingGLCalls(void)
   application.Render();
 
   DALI_TEST_EQUALS( testBox.x, glScissorParams.x, TEST_LOCATION );
-  DALI_TEST_EQUALS( testBox.y, stage.GetSize().height - glScissorParams.y - testBox.height, TEST_LOCATION ); // GL Coordinates are from bottom left
+  DALI_TEST_EQUALS( testBox.y, (int)(stage.GetSize().height - glScissorParams.y - testBox.height), TEST_LOCATION ); // GL Coordinates are from bottom left
   DALI_TEST_EQUALS( testBox.width, glScissorParams.width, TEST_LOCATION );
   DALI_TEST_EQUALS( testBox.height, glScissorParams.height, TEST_LOCATION );
   END_TEST;
